pyproject.toml
[project]
name = "plasmid-llm"
version = "0.1.0"
description = "Language models for plasmid DNA sequence generation"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1,<2.10",
    "transformers>=4.38",
    "tokenizers>=0.15",
    "pyarrow>=14.0",
    "pandas>=2.0",
    "mlflow>=2.10",
    "boto3>=1.28",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.2"]
rl = [
    "ray[default]>=2.9",     # Distributed post-training
    "parasail>=1.3.0",       # Smith-Waterman sequence alignment
    "biopython>=1.80",       # DNA translation utilities
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["plasmid_llm"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
